+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - The `go.opentelemetry.io/contrib/config` add support to configure periodic reader interval and timeout. (#5661)
 - Add support to configure views when creating MeterProvider using the config package. (#5654)
+- The `go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux` Add Hijack method to support middleware with websockets. (#5402)
 
 ### Fixed